src/api/python/requirements.txt
asgiref==3.4.1
click==8.0.1
fastapi==0.68.1
greenlet==1.1.2
h11==0.12.0
pydantic==1.8.2
pyodbc==4.0.32
python-dotenv==0.19.0
six==1.16.0
SQLAlchemy==1.4.25
SQLAlchemy-Utils==0.37.8
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
